ListViews  and User Interface Updated
ListViews are updated. Added check boxes next to every order in "Order Details" page to help keep track of the items bought and to be bought.  "Shop to Buy" and "Amount" details have been added to the list in the "Order Details" page. Prepared the application for merge and database connection. User Interface have been updated to work properly with more phones with different screen sizes.